+++ nodejs_container/native/Cargo.toml
@@ -26,3 +26,4 @@ holochain_core = { path = "../../core" }
 holochain_net = { path = "../../net" }
 holochain_core_types = { path = "../../core_types" }
 holochain_cas_implementations = { path = "../../cas_implementations" }
+snowflake = "1.2"
